+++ app/src/androidTest/java/com/harrysoft/burstcoinexplorer/test/BurstInfoServiceTest.java
@@ -4,8 +4,8 @@ import android.net.Uri;
 import android.support.test.InstrumentationRegistry;
 import android.support.test.runner.AndroidJUnit4;
 
-import com.harrysoft.burstcoinexplorer.burst.api.BurstInfoService;
-import com.harrysoft.burstcoinexplorer.burst.api.RepoInfoService;
+import com.harrysoft.burstcoinexplorer.burst.service.BurstInfoService;
+import com.harrysoft.burstcoinexplorer.burst.service.RepoInfoService;
 import com.harrysoft.burstcoinexplorer.burst.entity.EventInfo;
 import com.harrysoft.burstcoinexplorer.util.SingleTestUtils;
