+++ app/src/androidTest/java/com/harrysoft/burstcoinexplorer/test/BurstNetworkServiceTest.java
@@ -3,8 +3,8 @@ package com.harrysoft.burstcoinexplorer.test;
 import android.support.test.InstrumentationRegistry;
 import android.support.test.runner.AndroidJUnit4;
 
-import com.harrysoft.burstcoinexplorer.burst.api.BurstNetworkService;
-import com.harrysoft.burstcoinexplorer.burst.api.PoCCNetworkService;
+import com.harrysoft.burstcoinexplorer.burst.service.BurstNetworkService;
+import com.harrysoft.burstcoinexplorer.burst.service.PoCCNetworkService;
 import com.harrysoft.burstcoinexplorer.burst.entity.NetworkStatus;
 import com.harrysoft.burstcoinexplorer.util.SingleTestUtils;
