+++ app/src/androidTest/java/com/harrysoft/burstcoinexplorer/test/DetermineSearchRequestTypeTest.java
@@ -3,8 +3,8 @@ package com.harrysoft.burstcoinexplorer.test;
 import android.support.test.InstrumentationRegistry;
 
 import com.harrysoft.burstcoinexplorer.TestVariables;
-import com.harrysoft.burstcoinexplorer.burst.api.BurstBlockchainService;
-import com.harrysoft.burstcoinexplorer.burst.api.PoCCBlockchainService;
+import com.harrysoft.burstcoinexplorer.burst.service.BurstBlockchainService;
+import com.harrysoft.burstcoinexplorer.burst.service.PoCCBlockchainService;
 import com.harrysoft.burstcoinexplorer.burst.entity.SearchRequestType;
 import com.harrysoft.burstcoinexplorer.util.SingleTestUtils;
