+++ app/src/main/java/com/harrysoft/burstcoinexplorer/explore/browse/ViewBlockExtraDetailsActivity.java
@@ -6,7 +6,7 @@ import android.widget.TextView;
 import android.widget.Toast;
 
 import com.harrysoft.burstcoinexplorer.R;
-import com.harrysoft.burstcoinexplorer.burst.api.BurstBlockchainService;
+import com.harrysoft.burstcoinexplorer.burst.service.BurstBlockchainService;
 import com.harrysoft.burstcoinexplorer.burst.entity.BlockExtra;
 
 import java.math.BigInteger;
