+++ app/src/main/java/com/harrysoft/burstcoinexplorer/explore/browse/ViewTransactionsActivity.java
@@ -7,7 +7,7 @@ import android.widget.AdapterView;
 import android.widget.TextView;
 
 import com.harrysoft.burstcoinexplorer.R;
-import com.harrysoft.burstcoinexplorer.burst.api.BurstBlockchainService;
+import com.harrysoft.burstcoinexplorer.burst.service.BurstBlockchainService;
 import com.harrysoft.burstcoinexplorer.router.ExplorerRouter;
 
 import java.math.BigInteger;
